Viewer tests & rearrange test structure
